Move conftest into a subfolder.
