+++ tests/conftest.py
@@ -22,6 +22,8 @@ from platformSetup.firecracker import FireCracker
 from platformSetup.kvm import KVM
 from platformSetup.manual import Manual
 
+from conftest.architecture import *
+
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
@@ -506,27 +508,6 @@ def non_usi(testconfig):
         pytest.skip(f"test not supported on usi")    
 
 
-####
-# Feature ARCHITECTUR
-####
-
-@pytest.fixture
-def non_amd64(client):
-    (exit_code, output, error) = client.execute_command("dpkg --print-architecture", quiet=True)
-    if exit_code != 0:
-        logger.error(error)
-        sys.exit(exit_code)
-    if "amd64" in output:
-        pytest.skip('test not supported on amd64 architecture')
-
-@pytest.fixture
-def non_arm64(client):
-    (exit_code, output, error) = client.execute_command("dpkg --print-architecture", quiet=True)
-    if exit_code != 0:
-        logger.error(error)
-        sys.exit(exit_code)
-    if "arm64" in output:
-        pytest.skip('test not supported on arm64 architecture')
 
 
 #######
